third_party/blink/web_tests/external/wpt/domparsing/XMLSerializer-serializeToString-expected.txt

This is a testharness.js-based test.
PASS check XMLSerializer.serializeToString method could parsing xmldoc to string
PASS Check if the default namespace is correctly reset.
PASS Check if there is no redundant empty namespace declaration.
PASS check XMLSerializer.serializeToString escapes attribute values for roundtripping
PASS Check if generated prefixes match to "ns${index}".
FAIL Check if "ns1" is generated even if the element already has xmlns:ns1. assert_equals: expected "<root xmlns:ns2=\"uri2\"><child xmlns:ns1=\"uri1\" xmlns:ns1=\"uri3\" ns1:attr1=\"value1\"/></root>" but got "<root xmlns:ns2=\"uri2\"><child xmlns:ns1=\"uri1\" xmlns:ns3=\"uri3\" ns3:attr1=\"value1\"/></root>"
Harness: the test ran to completion.
